+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Use proxyequal for deep change detection
 
 ## [0.1.0] - 2018-11-15
 ### Added